+++ .gitignore
@@ -4,3 +4,5 @@
 *.env
 .vs
 .sarif
+
+.idea/
