+++ lf-sample-csharp-script-rules.sln
@@ -3,9 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "lf-sample-csharp-script-rules", "sample-rules\lf-sample-csharp-script-rules.csproj", "{CA916DC4-E454-441A-8821-E892CC73AF0A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "lf-sample-csharp-script-rules", "src\sample-rules\lf-sample-csharp-script-rules.csproj", "{CA916DC4-E454-441A-8821-E892CC73AF0A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "lf-sample-csharp-script-rules-test", "sample-rules-test\lf-sample-csharp-script-rules-test.csproj", "{EB135131-9566-449D-8D4A-FA170641E736}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "lf-sample-csharp-script-rules-test", "test\sample-rules-test\lf-sample-csharp-script-rules-test.csproj", "{EB135131-9566-449D-8D4A-FA170641E736}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{C0493191-6DEF-47ED-8A6C-C547DB2D52A3}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{1CF41641-2DF0-4B06-9411-0E9FEC92EA42}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,5 +31,9 @@ Global
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6BF86448-FC9F-427C-8166-CF47AC0F27F9}
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{CA916DC4-E454-441A-8821-E892CC73AF0A} = {C0493191-6DEF-47ED-8A6C-C547DB2D52A3}
+		{EB135131-9566-449D-8D4A-FA170641E736} = {1CF41641-2DF0-4B06-9411-0E9FEC92EA42}
 	EndGlobalSection
 EndGlobal
